LayoutTests/imported/w3c/web-platform-tests/webmessaging/postMessage_origin_mismatch.sub-expected.txt

CONSOLE MESSAGE: Unable to post message to http://www.invalid-domain.com. Recipient has origin http://localhost:8800.

CONSOLE MESSAGE: Unable to post message to http://www1.localhost:8800. Recipient has origin http://localhost:8800.


PASS Test Description: Same-origin: If the origin of the target window doesn't match the given origin, the message is discarded.
